添加Api请求接口方法

+++ app/src/main/java/com/pupu/rushui/entity/WhiteNoise.java
@@ -30,7 +30,7 @@ public class WhiteNoise extends BaseDO {
     /**
      * 键值
      */
-    int id;
+    Long id;
 
     /**
      * 下载地址
@@ -64,11 +64,11 @@ public class WhiteNoise extends BaseDO {
         this.name = whiteNoiseName;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
